visualization_flowchart/requirements.txt
requests>=2.31.0
pymupdf
python-dotenv
pydantic
loguru
fpdf2
Pillow